Able to run calculation
